import modifications across all folders
