+++ server/routes/BlogRoutes.js
@@ -5,7 +5,11 @@ const express = require("express");
 
 const prisma = new PrismaClient();
 const router = express.Router();
+const app = express();
+const blogRouter = require("./routes/blog.router");
 
+app.use(express.json());
+app.use(blogRouter);
 //// GET METHODS ////
 
 // Get all blogs
